refactor: moved bhandara-checkin to its respective folder
