+++ src/index.tsx
@@ -1,5 +1,5 @@
 import { render } from "react-dom";
-import { BhandaraCheckinWidgetDemo } from "./widgets/BhandaraCheckin/BhandaraCheckinDemo";
+import { BhandaraCheckinWidgetDemo } from "./widgets/BhandaraCheckin/BhandaraCheckin/BhandaraCheckinDemo";
 
 const root = document.getElementById("root");
