+++ src/widgets/BhandaraCheckin/BhandaraCheckin/BhandaraCheckinDemo.tsx
@@ -1,5 +1,5 @@
 import { BhandaraCheckinWidget } from "./BhandaraCheckinWidget";
-import { mockedApis } from "./mocked-apis";
+import { mockedApis } from "../mocked-apis";
 
 export const BhandaraCheckinWidgetDemo = () => {
   return <BhandaraCheckinWidget apis={mockedApis} />;
