+++ src/widgets/BhandaraCheckin/BhandaraCheckin/bhandara-checkin.ts
@@ -1,6 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { CurrentSectionEnum } from "../../types";
-import { BhandaraCheckinViewStateProps } from "../../BhandaraCheckinView";
+import { CurrentSectionEnum } from "../types";
+import { BhandaraCheckinViewStateProps } from "./BhandaraCheckinView";
 
 export const getBhandaraCheckinInitialState =
   (): BhandaraCheckinViewStateProps => {
