+++ src/widgets/BhandaraCheckin/store/slices/index.ts
@@ -1,3 +1,3 @@
-export * from "./bhandara-checkin";
+export * from "../../BhandaraCheckin/bhandara-checkin";
 export * from "../../SectionMain/mainSectionSlice";
 export * from "../../SectionUpdateDetails/updateDetailsSectionSlice";
